IR: Separate node types into sub-package
