+++ loki/ir/nodes/__init__.py
@@ -5,14 +5,8 @@
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
 
-"""
-The Loki internal representation (IR) and associated APIs for tree traversal.
-"""
+""" The node defininition classes for the Loki IR. """
 
-from loki.ir.expr_visitors import *  # noqa
-from loki.ir.find import *  # noqa
-from loki.ir.ir_graph import *  # noqa
-from loki.ir.nodes import *  # noqa
-from loki.ir.pragma_utils import *  # noqa
-from loki.ir.transformer import *  # noqa
-from loki.ir.visitor import *  # noqa
+from loki.ir.nodes.abstract_nodes import *  # noqaOA
+from loki.ir.nodes.internal_nodes import *  # noqa
+from loki.ir.nodes.leaf_nodes import *  # noqa
